requirements.txt
pytest>=7.0
python-dotenv>=1.0
Flask>=2.0
cryptography>=45.0
requests>=2.32
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp-proto-grpc>=1.27.0
opentelemetry-instrumentation-flask>=0.48b0
opentelemetry-instrumentation-requests>=0.48b0
